Cancel in UI, fixed deadlock in racing condition.

+++ action_server/tests/action_server_tests/test_server_async/async_actions.py
@@ -2,7 +2,7 @@ from sema4ai.actions import action
 
 
 @action
-def sleep_action(time_to_sleep: int) -> int:
+def sleep_action(time_to_sleep: float) -> float:
     """
     Args:
         time_to_sleep: The number of seconds to sleep.
